requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==3.2.10
django-allauth==0.47.0
django-allauth-2fa==0.8
django-crispy-forms==1.13.0
django-otp==1.1.3
idna==3.3
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
qrcode==7.3.1
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
urllib3==1.26.8
